Move component up
